+++ src/main/java/io/github/yoglappland/spectralization/optics/pump/OpticalPumpSource.java
@@ -1,5 +1,5 @@
 package io.github.yoglappland.spectralization.optics.pump;
 
 public interface OpticalPumpSource {
-    int pumpAmount();
+    double pumpAmount();
 }
